+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 0.0.7 - in progress
 - Support iteration over result sets.
 - Add python environments to gitignore.
+- Updating doc tests, adding new query type
+- Add new parameter to enable gene->dataset and protein->dataset queries
 - Running tests on client test server
 
 0.0.6 - 2021-03-24
@@ -14,8 +16,6 @@
 - Remove `values_type` parameter
 - `ResultsList` supports `[]`
 - Generate pydocs, with accurate kwarg lists..
-- Updating doc tests, adding new query type
-- Add new parameter to enable gene->dataset and protein->dataset queries
 
 0.0.5 - 2021-02-06
 - Support queries for and by dataset